new: arty: fix for pin conflict
Pin V17 was assigned to two different non-exclusive interfaces. This causes the tools to fail if both interfaces are used in a design. This was introduced when we changed the SPI interface to use IO10 for the SS pin to increase shield compatibility. Unfortunately this created a conflict with the GPIO interface connected to IO0-IO19. To fix the conflict I assigned the GPIO pin to the location attached to the SS pin on the SPI connector. I will create an issue to find a better fix for this.