new: zybo-z7: remove unneeded presets
By removing these Zynq preset options that are not required for the Zybo Z7,
It is possbile to rerun block automation on the Zynq IP core in a project and
not have the tools overwrite any manual changes made to the core by the user.
This is very useful when migrating a project from one board to another. See the
Zybo Z7 migration guide for more info.

Other Zynq boards should also be "pruned" like this.